+++ data-filter/lib/decorators/count.decorator.ts
@@ -2,9 +2,12 @@ import { AttributesHandler } from "../handlers/attributes.handler";
 import { DataFilterHandler } from "../handlers/data-filter.handler";
 import { CountAttributesConfig, CountConfig } from "../models/count.model";
 
-export function Count(name: string, options: CountConfig): PropertyDecorator & ClassDecorator {
+export function Count(name: string, options: Omit<CountConfig, "name">): PropertyDecorator & ClassDecorator {
     return (target: Object, propertyKey?: string) => {
-        defineCustomAttributesMetadata(target, propertyKey, name, options);
+        defineCustomAttributesMetadata(target, propertyKey, name, {
+            ...options,
+            name
+        });
     };
 }