+++ data-filter/lib/decorators/paranoid.decorator.ts
@@ -4,7 +4,7 @@ import { PathConfig } from "../models/path.model";
 export function Paranoid(paranoid = true): PropertyDecorator {
     return (target: Object, propertyKey: string) => {
         const attribute = AttributesHandler.getAttribute(target, propertyKey);
-        const path = attribute.path ?? {} as PathConfig;
+        const path = attribute.path ?? new PathConfig();
         path.paranoid = paranoid;
         attribute.setPath(path);
         AttributesHandler.saveAttribute(target, attribute);