+++ data-filter/lib/decorators/sum.decorator.ts
@@ -2,9 +2,12 @@ import { AttributesHandler } from "../handlers/attributes.handler";
 import { DataFilterHandler } from "../handlers/data-filter.handler";
 import { SumAttributesConfig, SumConfig } from "../models/sum.model";
 
-export function Sum(name: string, options: SumConfig): PropertyDecorator & ClassDecorator {
+export function Sum(name: string, options: Omit<SumConfig, "name">): PropertyDecorator & ClassDecorator {
     return (target: Object, propertyKey?: string) => {
-        defineCustomAttributesMetadata(target, propertyKey, name, options);
+        defineCustomAttributesMetadata(target, propertyKey, name, {
+            ...options,
+            name
+        });
     };
 }