+++ data-filter/lib/models/custom-attributes.model.ts
@@ -1,14 +1,21 @@
 import { ProjectionAlias } from "sequelize";
 import { GroupOption } from "sequelize/types/model";
+import { IncludeWhereModel } from "./include.model";
 import { PathModel } from "./path.model";
 
-export interface CustomAttributesConfig<T = any> {
+export interface CustomAttributesOptionConfig {
+    name: string;
+    path?: string;
+    where?: IncludeWhereModel;
+}
+
+export interface CustomAttributesConfig<T extends CustomAttributesOptionConfig = CustomAttributesOptionConfig> {
     key: string;
     type: string;
     config?: T;
 
     transform(options?: object, path?: string): string | ProjectionAlias;
-    groupBy(): GroupOption;
+    shouldGroupBy(): boolean;
 }
 
 export interface CustomAttributesModel {