+++ data-filter/lib/models/path.model.ts
@@ -1,17 +1,6 @@
 import { Order, WhereOptions } from "sequelize";
+import { SequelizeUtils } from "../sequelize.utils";
 import { IncludeWhereModel } from "./include.model";
-
-export interface PathConfig {
-    path: string;
-    where?: IncludeWhereModel;
-    required?: boolean;
-    separate?: boolean;
-    paranoid?: boolean;
-    subQuery?: boolean;
-    order?: Order;
-    limit?: number;
-}
-
 
 export interface PathModel {
     path: string;
@@ -22,4 +11,35 @@ export interface PathModel {
     subQuery?: boolean;
     order?: Order;
     limit?: number;
+}
+
+export class PathConfig {
+    path: string;
+    where?: IncludeWhereModel;
+    required?: boolean;
+    separate?: boolean;
+    paranoid?: boolean;
+    subQuery?: boolean;
+    order?: Order;
+    limit?: number;
+
+    constructor(values?: Partial<PathConfig>) {
+        Object.assign(this, values ?? {});
+    }
+
+    public transformPathConfig(options?: object): PathModel {
+        if (!options) {
+            const { where, ...path } = this;
+            return path;
+        }
+
+        if (this.where) {
+            return {
+                ...this,
+                where: SequelizeUtils.generateWhereConditions(this.where, options)
+            };
+        }
+
+        return { ...this } as PathModel;
+    }
 }