+++ data-filter/lib/models/sum.model.ts
@@ -1,11 +1,11 @@
+import { tr } from "date-fns/locale";
 import { fn, literal, ProjectionAlias } from "sequelize";
 import { GroupOption } from "sequelize/types/model";
 import { SequelizeUtils } from "../sequelize.utils";
-import { CustomAttributesConfig } from "./custom-attributes.model";
+import { CustomAttributesConfig, CustomAttributesOptionConfig } from "./custom-attributes.model";
 
-export interface SumConfig {
+export interface SumConfig extends CustomAttributesOptionConfig {
     attribute: string;
-    path?: string;
 }
 
 export class SumAttributesConfig implements CustomAttributesConfig<SumConfig> {
@@ -20,11 +20,7 @@ export class SumAttributesConfig implements CustomAttributesConfig<SumConfig> {
         return [fn("SUM", attribute), this.key];
     }
 
-    public groupBy(): GroupOption {
-        const attribute = this.config.path
-            ? literal(SequelizeUtils.getLiteralFullName(this.config.attribute, this.config.path))
-            : this.config.attribute;
-
-        return [attribute as string];
+    public shouldGroupBy(): boolean {
+        return true;
     }
 }